feat(git-flow-init): Add support for a second develop branch
This update introduces the option to configure a second develop branch
during the git-flow initialization process. It checks for the existence
of the branch and prompts the user for a name if it doesn't exist,
ensuring that it differs from the master and primary develop branches.
Additionally, it creates the second develop branch if necessary,
enhancing the flexibility of the git-flow setup.